+++ src/pages/calendar/styles/index.css
@@ -1,24 +1,30 @@
 /* CalendarScreen.css */
 
-/* 방법 1: 더 구체적인 선택자 사용 */
+/* 1. Calendar 최상위 컴포넌트 style */
 .react-calendar {
   display: flex;
   flex-direction: column;
-
   margin-top: 72px;
   padding-left: 24px !important;
   padding-right: 24px !important;
   box-sizing: border-box;
 }
 
-/* 네비게이션 바 레이아웃 변경 */
+/* 2. Calendar를 구성하는 모든 컴포넌트의 테두리 제거 및 폰트 스타일 적용*/
+.react-calendar * {
+  border: none;
+  font-family: var(--ff-sans);
+}
+
+/* 3. 네비게이션 바 레이아웃 변경 */
 .react-calendar__navigation {
   display: flex !important;
   justify-content: flex-start !important;
   align-items: center !important;
+  padding: 11.5px 0;
 }
 
-/* 날짜 레이블을 왼쪽에 배치 */
+/* 3.1. 네비게이션 바의 날짜 레이블 왼쪽에 배치 */
 .react-calendar__navigation__label {
   background: none;
   order: 1 !important;
@@ -27,13 +33,14 @@
   padding: 0;
 }
 
+/* 3.2. 네비게이션 바의 날짜 레이블 텍스트 스타일 변경 */
 .react-calendar__navigation__label__labelText {
   display: flex;
   font-size: 24px;
-  font-weight: 700px;
+  font-weight: 700;
 }
 
-/* 네비게이션 버튼들을 오른쪽에 배치 */
+/* 3.3. 네비게이션 화살표 버튼들을 오른쪽에 배치 */
 .react-calendar__navigation__prev-button,
 .react-calendar__navigation__next-button,
 .react-calendar__navigation__prev2-button,
@@ -44,8 +51,20 @@
   background: none;
 }
 
+.react-calendar__navigation__prev-button,
+.react-calendar__navigation__next-button {
+  margin-bottom: 5px;
+  font-size: 24px;
+}
+
+/* */
+.react-calendar__viewContainer {
+  margin-top: 33px;
+}
+
 /* 그리드 레이아웃 고정 - 핵심 해결책 */
-.react-calendar__month-view__days {
+.react-calendar__month-view__days,
+.react-calendar__month-view__weekdays {
   display: grid !important;
   grid-template-columns: repeat(7, 1fr) !important;
   grid-gap: 0 15px !important;
@@ -59,11 +78,19 @@
   flex-direction: column-reverse;
   justify-content: center;
   align-items: center;
-  padding: 0;
+  /* padding: 0; */
 }
 
-.react-calendar__month-view__weekdays > .react-calendar__month-view__weekdays__weekday {
-  text-align: center;
+.react-calendar__month-view__days__day--neighboringMonth {
+  opacity: 0 !important;
+}
+
+/* 요일 헤더 스타일 정렬 */
+.react-calendar__month-view__weekdays__weekday {
+  min-width: 0 !important;
+  width: 100% !important;
+  box-sizing: border-box !important;
+  text-align: center !important;
 }
 
 .react-calendar__month-view__weekdays > .react-calendar__month-view__weekdays__weekday > abbr {
@@ -71,8 +98,29 @@
   font-size: 12px;
 }
 
+.react-calendar__month-view__days__day {
+  transition: 0.2s background-color linear;
+  border-radius: 10px;
+  padding: 9px 0;
+  position: relative;
+}
+
+.react-calendar__month-view__days__day:hover {
+  background-color: #f0f0f0;
+}
+
+.react-calendar__month-view__days__day:focus > abbr {
+  background-color: var(--natural-400);
+}
+
 .react-calendar__month-view__days__day > abbr {
+  display: flex;
+  justify-content: center;
+  align-items: center;
   font-size: 12px;
+  width: 20px;
+  height: 20px;
+  border-radius: 50%;
 }
 
 .react-calendar__navigation__prev2-button,
